feat: add errors
