+++ src/types/IUser.ts
@@ -1,0 +1,8 @@
+export interface IUser {
+  id: string;
+  username: string;
+  age: number;
+  hobbies: string[]
+}
+
+export type DBtype = IUser[];
