+++ src/types/Methods.ts
@@ -1,0 +1,6 @@
+export enum Methods {
+  GET = 'GET',
+  POST = 'POST',
+  PUT = 'PUT',
+  DELETE = 'DELETE'
+}
