+++ src/types/StatusCodes.ts
@@ -1,0 +1,8 @@
+export enum StatusCodes {
+  OK = 200,
+  Created = 201,
+  NoContent = 204,
+  BadRequest = 400,
+  NotFound = 404,
+  InternalServerError = 500
+}
