+++ src/wrongRoute.ts
@@ -1,6 +1,7 @@
 import { ServerResponse } from "http"
+import { StatusCodes } from "./types/StatusCodes"
 
 export const wrongRoute = (response: ServerResponse) => {
-  response.writeHead(404, { 'Content-Type': 'application/json'})
+  response.writeHead(StatusCodes.NotFound, { 'Content-Type': 'application/json'})
   response.end(JSON.stringify({ message: 'Route not found' }))
 }
